+++ tests/testthat/test-prioritize.R
@@ -216,3 +216,14 @@ test_that("passes if score=1 & values by id_loan are duplicated for distinct
 
   expect_no_error(prioritize(valid))
 })
+
+test_that("with 0-row input returns 0-row input", {
+  lbk <-fake_lbk()
+  ald <- fake_ald(name_company = "won't match")
+  cero_row <- suppressWarnings(match_name(lbk, ald))
+
+  has_cero_row <- identical(nrow(cero_row), 0L)
+  stopifnot(has_cero_row)
+
+  expect_no_error(prioritize(cero_row))
+})